+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/models/_models_py3.py
@@ -4504,7 +4504,6 @@ class Machine(ProxyResource):
         "name": {"readonly": True},
         "type": {"readonly": True},
         "system_data": {"readonly": True},
-        "zones": {"readonly": True},
     }
 
     _attribute_map = {
@@ -4516,14 +4515,14 @@ class Machine(ProxyResource):
         "zones": {"key": "zones", "type": "[str]"},
     }
 
-    def __init__(self, *, properties: Optional["_models.MachineProperties"] = None, **kwargs: Any) -> None:
+    def __init__(self, *, properties: Optional["_models.MachineProperties"] = None, zones: Optional[List[str]] = None, **kwargs: Any) -> None:
         """
         :keyword properties: The properties of the machine.
         :paramtype properties: ~azure.mgmt.containerservice.models.MachineProperties
         """
         super().__init__(**kwargs)
         self.properties = properties
-        self.zones: Optional[List[str]] = None
+        self.zones = zones
 
 
 class MachineBillingProfile(_serialization.Model):